virtual timer

+++ source/app/loggerMetadata/loggerMetadata.hpp
@@ -41,11 +41,11 @@ constexpr uint16_t loggerNameLenght = 96;
  */
 struct loggerMetadata
 {
-	char	loggerName[loggerNameLenght];
-	uint8_t fileCreationPeriod = 4;	  // loggerSubsystem has to create files, this parameter indicates how often it does. (for more information, read externalStorage_component.hpp documentation)
-	uint8_t fileTransmissionPeriod;	  // Period (minutes) for sending the current file to the server
-	uint8_t generalMeasurementPeriod; // Period (seconds) for making a measurement and storing it (not all sensors follow this period)
-	uint8_t restRequestPeriod;		  // Period (minutes) for sending last computed data line to the server
+	char	 loggerName[loggerNameLenght];
+	uint16_t fileCreationPeriod = 4;   // loggerSubsystem has to create files, this parameter indicates how often it does. (for more information, read externalStorage_component.hpp documentation)
+	uint16_t fileTransmissionPeriod;   // Period (minutes) for sending the current file to the server
+	uint16_t generalMeasurementPeriod; // Period (seconds) for making a measurement and storing it (not all sensors follow this period)
+	uint16_t restRequestPeriod;		   // Period (minutes) for sending last computed data line to the server
 };
 
 ////////////////////////////////////////////////////////////////////////
